Style Changes
moved the save Tree button into the buttons div along with the new member and new relation buttons.
They are now all located in the sidebar.
I changed the styling of the side bar so that it collapses when the form is not there and the buttons appear and disappear opposite of when the form does.

in edit.py I removed the function that referenced the members.html page because we are no longer going to be doing that.

and other small css tweaks.

Delete saveButton.png