Fixed some bugs
Uploading to github messed with some of my files this is the most updated fully working code

Bug Fixes

Changed members and relations to have unique ID's. This fixed some bugs that we're introduced when I switched the relations from referencing the member index instead of the member object.

I removed the "Home" button from the profile page.

I added a save tree button image and put the reference in the spec. document

I had to change the save and load to handle the relation and member IDs

I also added arrow key movement controls