fixed dotted relations not saving
also removed the help button JQuery stuff because the help button does not exist anymore and there was an error every time you opened the edit page because of it.

Changed the relations so that they have a mode 'h' or 'v' which allows them to be draw horizontally or vertically. and there is now a option in the sidebar to toggle that instead of using ' , '